# public/swagger/v2/api.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GC2 API",
        "contact": {
            "email": "mh@mapcentia.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v2/configuration/{userId}/{configurationId}": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Returns specific configuration (unpublished configuration is returned if authorized user is the author)",
                "operationId": "app\\api\\v2\\Configuration::get_specific",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "configurationId",
                        "in": "path",
                        "description": "Configuration id (key)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Deletes configuration",
                "operationId": "app\\api\\v2\\Configuration::delete_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "configurationId",
                        "in": "path",
                        "description": "Configuration id (key)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/v2/configuration/{userId}/": {
            "post": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Creates configuration",
                "operationId": "app\\api\\v2\\Configuration::post_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Configuration data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "published": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/apiv2/configuration/{userId}/{configurationId}": {
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Creates configuration",
                "operationId": "app\\api\\v2\\Configuration::put_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "configurationId",
                        "in": "path",
                        "description": "Configuration id (key)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Configuration data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "published": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
        "/api/v2/keyvalue/{userId}/{key}": {
            "get": {
                "tags": [
                    "Keyvalue"
                ],
                "summary": "Get value by key",
                "operationId": "app\\api\\v2\\Keyvalue::get_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key to fetch",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Keyvalue"
                ],
                "summary": "Update value in store",
                "operationId": "app\\api\\v2\\Keyvalue::put_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "New JSON value to store",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        },
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Keyvalue"
                ],
                "summary": "Create new key/value in store",
                "operationId": "app\\api\\v2\\Keyvalue::post_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key to create",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON value to store",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        },
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Keyvalue"
                ],
                "summary": "Delete a key/bvalue in store",
                "operationId": "app\\api\\v2\\Keyvalue::delete_index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "my_key"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v2/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Creates user",
                "operationId": "app\\api\\v2\\User::post_index",
                "requestBody": {
                    "description": "User data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234Luggage"
                                    },
                                    "subuser": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "usergroup": {
                                        "type": "string",
                                        "example": "My group"
                                    },
                                    "parentdb": {
                                        "type": "string",
                                        "example": "mydatabase"
                                    },
                                    "properties": {
                                        "type": "object",
                                        "example": {
                                            "org": "Ajax Inc."
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
        "/api/v2/user/{userId}/subusers": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Returns subusers",
                "operationId": "app\\api\\v2\\User::get_subusers",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation status"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    }
}